Fix workbench startup on Windows
Qt5 abstracts platform specifics to plugins so startup requires
that the Qt plugins are locatable. A startup function is implemented
that is a no-op on platforms other than Windows.
Refs #20381
