+++ buildconfig/windows/buildenv.bat.in
@@ -13,6 +13,9 @@ set THIRD_PARTY_DIR=%CM_THIRD_PARTY_DIR:/=\%
 :: Qt4 - exes are in bin & dlls are in lib
 set QT4_BIN=%THIRD_PARTY_DIR%\lib\qt4\bin;%THIRD_PARTY_DIR%\lib\qt4\lib
 
+:: Qt5 - exes are in bin & dlls are in lib
+set QT5_BIN=%THIRD_PARTY_DIR%\lib\qt5\bin;%THIRD_PARTY_DIR%\lib\qt5\lib
+
 :: Python -
 set PYTHONHOME=%THIRD_PARTY_DIR%\lib\python@PYTHON_MAJOR_VERSION@.@PYTHON_MINOR_VERSION@
 
@@ -20,4 +23,4 @@ set PYTHONHOME=%THIRD_PARTY_DIR%\lib\python@PYTHON_MAJOR_VERSION@.@PYTHON_MINOR_
 set MISC_BIN=%THIRD_PARTY_DIR%\bin;%THIRD_PARTY_DIR%\bin\mingw
 
 :: Update PATH
-set PATH=%MISC_BIN%;%PYTHONHOME%;%QT4_BIN%;%PATH%
+set PATH=%MISC_BIN%;%PYTHONHOME%;%QT5_BIN%;%QT4_BIN%;%PATH%
