+++ qt/python/.gitignore
@@ -1,3 +1,6 @@
+# File generated by the build
+mantidqt/utils/qt/plugins.py
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
